# agendamento/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Agendamento de Consultas",
    "version": "1.0.0",
    "description": "Documentação da API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/users/{userId}/specialties/{specialtyId}/appointments": {
      "get": {
        "summary": "Retorna consultas de um utilizador com médicos de uma especialidade específica",
        "description": "Lista todas as consultas agendadas para um utilizador cujo ID é especificado no caminho,\ncom médicos que pertencem à especialidade cujo ID também é especificado.\nAcessível apenas ao utilizador especificado (se autenticado) ou a um administrador.\n",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "O ID do utilizador (paciente) cujas consultas devem ser retornadas",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "specialtyId",
            "required": true,
            "description": "O ID da especialidade para filtrar os médicos associados às consultas",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de consultas encontrada para o utilizador e especialidade",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "O ID da consulta"
                      },
                      "date": {
                        "type": "string",
                        "format": "date",
                        "description": "A data da consulta"
                      },
                      "time": {
                        "type": "string",
                        "format": "time",
                        "description": "A hora da consulta"
                      },
                      "notes": {
                        "type": "string",
                        "description": "Notas da consulta (opcional)",
                        "nullable": true
                      },
                      "medico": {
                        "type": "object",
                        "description": "Os detalhes do médico associado à consulta",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "O ID do médico"
                          },
                          "name": {
                            "type": "string",
                            "description": "O nome do médico"
                          },
                          "specialty": {
                            "type": "object",
                            "description": "A especialidade do médico",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "description": "O ID da especialidade"
                              },
                              "name": {
                                "type": "string",
                                "description": "O nome da especialidade"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado. O token de autenticação está ausente ou inválido."
          },
          "403": {
            "description": "Acesso negado. O utilizador autenticado não tem permissão para aceder a estas consultas (não é o utilizador na URL nem admin)."
          },
          "404": {
            "description": "Utilizador ou especialidade não encontrados com os IDs fornecidos."
          },
          "500": {
            "description": "Erro interno do servidor ao processar a requisição."
          }
        }
      }
    },
    "/users/{userId}/appointments": {
      "get": {
        "summary": "Retorna todas as consultas de um utilizador",
        "description": "Lista todas as consultas agendadas para um utilizador cujo ID é especificado no caminho.\nAcessível apenas ao utilizador especificado (se autenticado) ou a um administrador.\n",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "O ID do utilizador cujas consultas devem ser retornadas",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de consultas encontrada para o utilizador",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "O ID da consulta"
                      },
                      "date": {
                        "type": "string",
                        "format": "date",
                        "description": "A data da consulta"
                      },
                      "time": {
                        "type": "string",
                        "format": "time",
                        "description": "A hora da consulta"
                      },
                      "notes": {
                        "type": "string",
                        "description": "Notas ou descrição da consulta (opcional)",
                        "nullable": true
                      },
                      "paciente": {
                        "type": "object",
                        "description": "Os detalhes do paciente (utilizador) associado a esta consulta",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "O ID do paciente"
                          },
                          "name": {
                            "type": "string",
                            "description": "O nome do paciente"
                          },
                          "email": {
                            "type": "string",
                            "format": "email",
                            "description": "O email do paciente"
                          }
                        }
                      },
                      "medico": {
                        "type": "object",
                        "description": "Os detalhes do médico associado a esta consulta",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "O ID do médico"
                          },
                          "name": {
                            "type": "string",
                            "description": "O nome do médico"
                          },
                          "specialty": {
                            "type": "object",
                            "description": "A especialidade do médico",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "description": "O ID da especialidade"
                              },
                              "name": {
                                "type": "string",
                                "description": "O nome da especialidade"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado. O token de autenticação está ausente ou inválido."
          },
          "403": {
            "description": "Acesso negado. O utilizador autenticado não tem permissão para aceder a estas consultas (não é o utilizador na URL nem admin)."
          },
          "404": {
            "description": "Utilizador não encontrado com o ID fornecido."
          },
          "500": {
            "description": "Erro interno do servidor ao processar a requisição."
          }
        }
      }
    },
...
    "/appointments": {
      "get": {
        "summary": "Retorna uma lista de consultas",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de consultas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "O ID da consulta"
                      },
                      "data": {
                        "type": "string",
                        "format": "date-time",
                        "description": "A data e hora da consulta"
                      },
                      "descricao": {
                        "type": "string",
                        "description": "Descrição da consulta (opcional)"
                      },
                      "specialty": {
                        "type": "object",
                        "description": "A especialidade da consulta",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "O ID da especialidade"
                          },
                          "name": {
                            "type": "string",
                            "description": "O nome da especialidade"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Cria uma nova consulta",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "data",
                  "pacienteId",
                  "medicoId"
                ],
                "properties": {
                  "data": {
                    "type": "string",
                    "format": "date-time",
                    "description": "A data e hora da consulta"
                  },
                  "descricao": {
                    "type": "string",
                    "description": "Descrição da consulta (opcional)"
                  },
                  "pacienteId": {
                    "type": "integer",
                    "description": "O ID do paciente relacionado"
                  },
                  "medicoId": {
                    "type": "integer",
                    "description": "O ID do médico relacionado"
                  },
                  "especialidadeId": {
                    "type": "integer",
                    "description": "O ID da especialidade desejada para a consulta"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consulta criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "descricao": {
                      "type": "string"
                    },
                    "pacienteId": {
                      "type": "integer"
                    },
                    "medicoId": {
                      "type": "integer"
                    },
                    "especialidadeId": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou ausentes"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  }
}